Divide change-signature test in two because throw max-line limit size
- Add new tests for modules for adding and removing parameter
