+++ src/refactorings/change-signature/change-signature.test.ts
@@ -6,11 +6,6 @@ import { testEach } from "../../tests-helpers";
 
 import { changeSignature, createVisitor } from "./change-signature";
 
-type TestSample = {
-  code: Code;
-  path: AbsolutePath;
-};
-
 describe("Change Signature", () => {
   testEach<{
     code: Code;
@@ -298,200 +293,6 @@ describe("Change Signature", () => {
       expect(extracted).toBe(expected);
     }
   );
-
-  describe("Modules", () => {
-    const addModule = new AbsolutePath("/temp/add.ts");
-    const anotherModule = new AbsolutePath("/temp/anotherModule");
-
-    testEach<{
-      setup: {
-        currentFile: TestSample;
-        otherFiles: TestSample[];
-      };
-      expected: {
-        currentFile: TestSample;
-        otherFiles: TestSample[];
-      };
-    }>(
-      "should change signature of function with a reference in modules",
-      [
-        {
-          description: "that import the function",
-          setup: {
-            currentFile: {
-              code: `export function [cursor]add(a, b) {
-              return a + b;
-            }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {add} from './module';
-                add(1, 2)
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  return add(a, b);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          },
-          expected: {
-            currentFile: {
-              code: `export function add(b, a) {
-              return a + b;
-            }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {add} from './module';
-                add(2, 1)
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  return add(b, a);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          }
-        },
-        {
-          description: "that import arrow function",
-          setup: {
-            currentFile: {
-              code: `export const add = [cursor](a, b) => {
-              return a + b;
-            }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {add} from './module';
-                add(1, 2)
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  return add(a, b);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          },
-          expected: {
-            currentFile: {
-              code: `export const add = (b, a) => {
-              return a + b;
-            }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {add} from './module';
-                add(2, 1)
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  return add(b, a);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          }
-        },
-        {
-          description: "that import a class and use a method",
-          setup: {
-            currentFile: {
-              code: `export class Maths {
-                [cursor]add(a, b) {
-                  return a + b;
-                }
-              }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {Maths} from './module';
-                const maths = new Maths();
-                maths.add(1, 2);
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  const maths = new Maths();
-                  return maths.add(a, b);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          },
-          expected: {
-            currentFile: {
-              code: `export class Maths {
-                add(b, a) {
-                  return a + b;
-                }
-              }`,
-              path: new AbsolutePath("/temp/module.ts")
-            },
-            otherFiles: [
-              {
-                code: `import {Maths} from './module';
-                const maths = new Maths();
-                maths.add(2, 1);
-              `,
-                path: addModule
-              },
-              {
-                code: `import {add} from './anotherModule';
-                export const calculateAdd = (a, b) => {
-                  const maths = new Maths();
-                  return maths.add(b, a);
-                }
-              `,
-                path: anotherModule
-              }
-            ]
-          }
-        }
-      ],
-      async ({ setup, expected }) => {
-        const editor = new InMemoryEditor(setup.currentFile.code);
-        await editor.writeIn(setup.currentFile.path, editor.code);
-        editor.saveUserChoices(userChangePositionOf(0, 1));
-        editor.saveUserChoices(userChangePositionOf(1, 0));
-        await saveOtherFiles(setup, editor);
-
-        await changeSignature(editor);
-
-        const extracted = await editor.codeOf(setup.currentFile.path);
-        expect(extracted).toBe(expected.currentFile.code);
-        await validateOutput(expected, editor);
-      }
-    );
-  });
 
   it("should show an error message if refactoring can't be made because rest param should be the last", async () => {
     const code = `
@@ -892,32 +693,6 @@ describe("Change Signature", () => {
     expect(extracted).toBe(expected);
   });
 });
-
-function validateOutput(
-  expected: { currentFile: TestSample; otherFiles: TestSample[] },
-  editor: InMemoryEditor
-) {
-  const promises = expected.otherFiles.map(async (file) => {
-    const extracted = await editor.codeOf(file.path);
-    expect(extracted).toBe(file.code);
-  });
-
-  return Promise.all(promises);
-}
-
-function saveOtherFiles(
-  setup: {
-    currentFile: TestSample;
-    otherFiles: TestSample[];
-  },
-  editor: InMemoryEditor
-) {
-  const promises = setup.otherFiles.map(async (file) => {
-    await editor.writeIn(file.path, file.code);
-  });
-
-  return Promise.all(promises);
-}
 
 function userChangePositionOf(
   startAt: number,
